backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
django-filter==24.3
django-celery-beat==2.7.0
django-storages==1.14.4
dj-database-url==2.3.0
psycopg2-binary==2.9.10
celery==5.4.0
redis==5.2.1
gunicorn==23.0.0
Pillow==11.0.0
python-decouple==3.8
requests==2.32.3
boto3==1.35.81
whitenoise==6.8.2
sentry-sdk==2.19.2
django-extensions==3.2.3
